% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_list2}
\alias{disease_list2}
\title{Disease List Dataset}
\format{
A data frame with 123 rows (diseases) and 5 variables:
\describe{
  \item{disease_name}{Name of the disease as is mentioned on Empres-i reports}
  \item{acronym}{English acronym for the disease}
  \item{portuguese_name}{The official portuguese name for the disease}
  \item{common_english_name}{Common English name for the disease}
  \item{portuguese_aconym}{Portuguese acronym for the disease}
  \item{link_EFSA}{Link for EFSA's page for the disease}
  \item{url}{link to WOAH's page for the disease}
}
}
\source{
\url{https://www.woah.org/}
}
\usage{
disease_list2
}
\description{
A dataset containing disease names alias.
}
\keyword{datasets}
