% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Empres.data}
\alias{Empres.data}
\title{Retrieve outbreak data}
\usage{
Empres.data(
  diseasename = NA,
  region = NA,
  country = NA,
  startdate = NA,
  enddate = NA,
  species = NA
)
}
\arguments{
\item{diseasename}{Disease name. To check valid names use disease_list()}

\item{region}{Globe region. Valid values are "Africa","Asia","Europe","Americas","Oceania"}

\item{country}{reporting country}

\item{species}{Affected species}

\item{start_date}{Starting observation date}

\item{end_date}{ending observation date}
}
\description{
Retrieve outbreak data
}
