% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/core.R
\name{Empres.data}
\alias{Empres.data}
\title{Retrieve outbreak data}
\usage{
Empres.data(disease = NA, region = NA, sodate = NA, eodate = NA,
  srdate = NA, erdate = NA)
}
\arguments{
\item{disease}{Disease name. To check valid names use disease_list()}

\item{region}{Globe region. Valid values are "Africa","Asia","Europe","Americas","Oceania"}

\item{sodate}{Starting observation date}

\item{eodate}{ending observation date}

\item{srdate}{Starting reporting date}

\item{erdate}{Ending reporting date}
}
\description{
Retrieve outbreak data
}

